requirements.txt
# PDF Processing
pypdf==4.0.1
pdf2image==1.17.0
Pillow==10.2.0

# Google Gemini API
google-generativeai==0.3.2

# Data Validation & Schema
pydantic==2.5.3
pydantic-settings==2.1.0
jsonschema==4.20.0

# XML Processing
lxml==5.1.0

# Document Generation
reportlab==4.0.9
python-docx==1.1.0
jinja2==3.1.3

# UI (Milestone 2)
streamlit==1.30.0

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.1
tenacity==8.2.3

# Logging
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==24.1.1
ruff==0.1.15
mypy==1.8.0
